only do it if there is an array there
git-svn-id: file:///home/oestrich/tmp/img/trunk@739 ea04b438-0a61-0410-93d0-9bc465a3bab0

+++ admin/queue.php
@@ -30,7 +30,7 @@ if($_GET['action'] == "update"){
 	}
 }
 
-if($_GET['action'] == "delete"){
+if($_GET['action'] == "delete" && is_array($_POST['approve'])){
 	foreach($_POST['approve'] as $key => $value){
 		$lncln->delete($key);
 	}
